+++ __tests__/fixtures/vba/src/queries/Consulta3.sql
@@ -1,0 +1,3 @@
+SELECT TbACParaLista.*
+FROM TbACParaLista
+ORDER BY FechaInicial, NAccion;